+++ GateAdventures/GateAdventures.cs
@@ -1,67 +1,69 @@
-﻿using System.Collections.Generic;
-using System.IO;
-using GateAdventures.Engine.Map;
+﻿using GateAdventures.Engine.Scenes;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Input;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 
-namespace GateAdventures
+namespace GateAdventures;
+
+public class GateAdventures : Game
 {
-	public class GateAdventures : Game
+	private readonly GraphicsDeviceManager _graphics;
+	private readonly SceneManager _sceneManager;
+
+
+	public GateAdventures()
 	{
-		private GraphicsDeviceManager _graphics;
-		private SpriteBatch _spriteBatch;
+		_graphics = new GraphicsDeviceManager(this);
+		_sceneManager = new SceneManager(Services);
 
-		private List<TileMapRenderer> _maps;
+		Content.RootDirectory = "Content";
+		IsMouseVisible = true;
+	}
 
-		public GateAdventures()
+	protected override void Initialize()
+	{
+		_graphics.PreferredBackBufferWidth = 960;
+		_graphics.PreferredBackBufferHeight = 960;
+		_graphics.ApplyChanges();
+
+		SpriteBatch spriteBatch = new(GraphicsDevice);
+
+		Services.AddService(spriteBatch);
+		Services.AddService(_graphics);
+		Services.AddService(GraphicsDevice);
+		Services.AddService(_sceneManager);
+		Services.AddService(Content);
+
+		_sceneManager.Start(new TileScene());
+		_sceneManager.StartOverlay(new HelloScene());
+
+		base.Initialize();
+	}
+
+	protected override void Update(GameTime gameTime)
+	{
+		IScene[] scenes = _sceneManager.Scenes;
+		for (int i = 0; i < scenes.Length; i++)
 		{
-			_graphics = new GraphicsDeviceManager(this);
-			Content.RootDirectory = "Content";
-			IsMouseVisible = true;
-		}
-
-		protected override void Initialize()
-		{
-			// TileMap width/height. We'll need to change this to 1920x1080
-			_graphics.PreferredBackBufferWidth = 960;
-			_graphics.PreferredBackBufferHeight = 960;
-
-			_graphics.ApplyChanges();
-			base.Initialize();
-		}
-
-		protected override void LoadContent()
-		{
-			_spriteBatch = new SpriteBatch(GraphicsDevice);
-
-			// Load all configured TileMaps and their corresponding TileSets
-			_maps = TiledMapLoader.LoadAllMaps(Content, _graphics.GraphicsDevice);
-		}
-
-		protected override void Update(GameTime gameTime)
-		{
-			if (GamePad.GetState(PlayerIndex.One).Buttons.Back == ButtonState.Pressed || Keyboard.GetState().IsKeyDown(Keys.Escape))
-				Exit();
-
-			base.Update(gameTime);
-		}
-
-		protected override void Draw(GameTime gameTime)
-		{
-			GraphicsDevice.Clear(Color.CornflowerBlue);
-			_spriteBatch.Begin();
-
-			// TODO: associate a TileMap with a scene and draw it there, rather than looping through and drawing all tilemaps
-			foreach (TileMapRenderer mapRenderer in _maps)
+			IScene scene = scenes[i];
+			if (i == (scenes.Length - 1))
 			{
-				mapRenderer.Draw(_spriteBatch);
+				scene.HandleInput();
 			}
 
-			_spriteBatch.End();
-			base.Draw(gameTime);
+			scene.Update(gameTime);
 		}
+
+		base.Update(gameTime);
+	}
+
+	protected override void Draw(GameTime gameTime)
+	{
+		IScene[] scenes = _sceneManager.Scenes;
+		foreach (IScene scene in scenes)
+		{
+			scene.Draw(gameTime);
+		}
+
+		base.Draw(gameTime);
 	}
 }